+++ src/main/java/graphicsEngineExtension/input/mouse/SimpleMouseListener.java
@@ -1,4 +1,4 @@
-package graphicsEngineExtension.mouse;
+package graphicsEngineExtension.input.mouse;
 
 import java.awt.event.MouseListener;
 import java.awt.event.MouseEvent;
